refactor(captp): breakout isOnlyLocal from engine
